Bug 2070400 - Pt 5 - Make PictureIndex a u32 rather than a usize r=gfx-reviewers,nical
It was the only thing forcing eight-byte alignment on PrimitiveKind, which
drops from 24 bytes to 16, and PrimitiveInstance from 32 to 24. SpatialNodeIndex
and ClipTaskIndex are already u32.

Differential Revision: https://phabricator.services.mozilla.com/D324498

+++ gfx/wr/webrender/src/tile_cache/slice_builder.rs
@@ -691,7 +691,7 @@ fn create_tile_cache(
         None,
     ));
 
-    tile_cache_pictures.push(PictureIndex(pic_index));
+    tile_cache_pictures.push(PictureIndex(pic_index as u32));
 }
 
 /// Debug information about a set of picture cache slices, exposed via RenderResults
